# server/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["登录验证"]}},"/api/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["用户详情"],"security":[{"bearer":[]}]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["用户详情"],"security":[{"bearer":[]}]}},"/api/user/info":{"get":{"operationId":"UserController_getUserInfo","parameters":[],"responses":{"200":{"description":""}},"tags":["用户详情"],"security":[{"bearer":[]}]}},"/api/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["用户详情"],"security":[{"bearer":[]}]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["用户详情"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["用户详情"],"security":[{"bearer":[]}]}},"/api/upload/file":{"post":{"operationId":"UploadController_uploadFile","parameters":[],"responses":{"201":{"description":""}},"tags":["文件上传"]}},"/api/upload/files":{"post":{"operationId":"UploadController_uploadFiles","parameters":[],"responses":{"201":{"description":""}},"tags":["文件上传"]}},"/api/static/{subPath}":{"get":{"operationId":"StaticController_render","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/quote/random":{"get":{"operationId":"QuoteController_getRandomQuote","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteDto"}}}}},"tags":["名言名句"]}},"/api/about":{"post":{"operationId":"AboutController_create","summary":"增加","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAboutDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["关于我们"],"security":[{"bearer":[]}]},"get":{"operationId":"AboutController_findAll","summary":"列表","parameters":[{"name":"keyword","required":false,"in":"query","schema":{"type":"string"}},{"name":"category","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""},"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/About"}}}}}},"tags":["关于我们"],"security":[{"bearer":[]}]}},"/api/about/{id}":{"get":{"operationId":"AboutController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/About"}}}}},"tags":["关于我们"],"security":[{"bearer":[]}]},"patch":{"operationId":"AboutController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAboutDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["关于我们"],"security":[{"bearer":[]}]},"delete":{"operationId":"AboutController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["关于我们"],"security":[{"bearer":[]}]}},"/api/about/list/count":{"get":{"operationId":"AboutController_getCount","summary":"用户数量","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["关于我们"],"security":[{"bearer":[]}]}},"/api/banner":{"post":{"operationId":"BannerController_create","summary":"增加","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBannerDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Banner"],"security":[{"bearer":[]}]},"get":{"operationId":"BannerController_findAll","summary":"列表","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}},"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Banner"}}}}}},"tags":["Banner"],"security":[{"bearer":[]}]}},"/api/banner/{id}":{"get":{"operationId":"BannerController_findOne","summary":"根据 id 查找","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Banner"}}}}},"tags":["Banner"],"security":[{"bearer":[]}]},"patch":{"operationId":"BannerController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBannerDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Banner"],"security":[{"bearer":[]}]},"delete":{"operationId":"BannerController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Banner"],"security":[{"bearer":[]}]}},"/api/category":{"post":{"operationId":"CategoryController_create","summary":"增加","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}}}},"tags":["category"]},"get":{"operationId":"CategoryController_findAll","summary":"列表","parameters":[{"name":"keyword","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["category"]}},"/api/category/{id}":{"get":{"operationId":"CategoryController_findOne","summary":"根据 id 查找","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}}}},"tags":["category"]},"patch":{"operationId":"CategoryController_update","summary":"更新","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":""}},"tags":["category"]},"delete":{"operationId":"CategoryController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["category"]}},"/api/links":{"post":{"operationId":"LinksController_create","summary":"增加","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLinkDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Link"}}}}},"tags":["links"]},"get":{"operationId":"LinksController_findAll","summary":"列表","parameters":[{"name":"keyword","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["links"]}},"/api/links/{id}":{"get":{"operationId":"LinksController_findOne","summary":"根据 id 查找","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Link"}}}}},"tags":["links"]},"patch":{"operationId":"LinksController_update","summary":"更新","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLinkDto"}}}},"responses":{"200":{"description":""}},"tags":["links"]},"delete":{"operationId":"LinksController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["links"]}}},"info":{"title":"nest admin boilerplate","description":"nest admin boilerplate 的 API 文档","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","maxLength":14,"description":"用户名"},"password":{"type":"string","description":"密码"},"email":{"type":"string","description":"邮箱"}},"required":["username","password"]},"User":{"type":"object","properties":{"id":{"type":"number","description":"自增 id"},"username":{"type":"string","description":"标题"},"password":{"type":"string","description":"密码"},"email":{"type":"string","description":"邮箱"},"is_admin":{"type":"number","description":"是否为管理员"}},"required":["id","username","password","email"]},"UpdateUserDto":{"type":"object","properties":{"username":{"type":"string","maxLength":14,"description":"用户名"},"password":{"type":"string","description":"密码"},"email":{"type":"string","description":"邮箱"}}},"QuoteDto":{"type":"object","properties":{"_id":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"content":{"type":"string"},"author":{"type":"string"},"authorSlug":{"type":"string"},"length":{"type":"number"},"dateAdded":{"type":"string"},"dateModified":{"type":"string"}},"required":["_id","tags","content","author","authorSlug","length","dateAdded","dateModified"]},"CreateAboutDto":{"type":"object","properties":{"name":{"type":"string"},"route":{"type":"string"},"description":{"type":"string"},"category":{"type":"string"},"banner":{"type":"string"},"content":{"type":"string"},"status":{"type":"boolean"},"createdAt":{"format":"date-time","type":"string"},"id":{"type":"number"},"updatedAt":{"format":"date-time","type":"string"}},"required":["name","route","description","category","banner","content","status","createdAt","id","updatedAt"]},"About":{"type":"object","properties":{"id":{"type":"number","description":"自增 id"},"name":{"type":"string","description":"name"},"route":{"type":"string","description":"route"},"description":{"type":"string","description":"description"},"category":{"type":"string","description":"category"},"banner":{"type":"string","description":"banner"},"content":{"type":"string","description":"content"},"status":{"type":"boolean","description":"status"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","name","route","description","category","banner","content","status","createdAt","updatedAt","deletedAt"]},"UpdateAboutDto":{"type":"object","properties":{"name":{"type":"string"},"route":{"type":"string"},"description":{"type":"string"},"category":{"type":"string"},"banner":{"type":"string"},"content":{"type":"string"},"status":{"type":"boolean"},"updatedAt":{"format":"date-time","type":"string"}},"required":["name","route","description","category","banner","content","status","updatedAt"]},"CreateBannerDto":{"type":"object","properties":{"name":{"type":"string"},"url":{"type":"string"},"intro":{"type":"string"},"status":{"type":"boolean"},"createdAt":{"format":"date-time","type":"string"},"id":{"type":"number"},"updatedAt":{"format":"date-time","type":"string"}},"required":["name","url","intro","status","createdAt","id","updatedAt"]},"Banner":{"type":"object","properties":{"id":{"type":"number","description":"自增 id"},"name":{"type":"string","description":"name"},"url":{"type":"string","description":"url"},"pic":{"type":"string","description":"pic"},"intro":{"type":"string","description":"intro"},"status":{"type":"boolean","description":"status"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","name","url","pic","intro","status","createdAt","updatedAt","deletedAt"]},"UpdateBannerDto":{"type":"object","properties":{"name":{"type":"string"},"url":{"type":"string"},"intro":{"type":"string"},"status":{"type":"boolean"},"updatedAt":{"format":"date-time","type":"string"}},"required":["name","url","intro","status","updatedAt"]},"CreateCategoryDto":{"type":"object","properties":{"name":{"type":"string","maxLength":80,"description":"category name"}},"required":["name"]},"Link":{"type":"object","properties":{"id":{"type":"number","description":"自增 id"},"name":{"type":"string","description":"name"},"title":{"type":"string","description":"title"},"description":{"type":"string","description":"description"},"isDelete":{"type":"boolean","description":"status"},"url":{"type":"string"},"icon":{"type":"string"},"weight":{"type":"number"},"category":{"$ref":"#/components/schemas/Category"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","name","title","description","isDelete","url","icon","weight","category","createdAt","updatedAt","deletedAt"]},"Category":{"type":"object","properties":{"id":{"type":"number","description":"自增 id"},"name":{"type":"string","description":"title"},"isDelete":{"type":"boolean","description":"status"},"parent":{"$ref":"#/components/schemas/Category"},"children":{"type":"array","items":{"$ref":"#/components/schemas/Category"}},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","name","isDelete","parent","children","links","createdAt","updatedAt","deletedAt"]},"UpdateCategoryDto":{"type":"object","properties":{"name":{"type":"string","maxLength":80,"description":"category name"},"updatedAt":{"format":"date-time","type":"string"}},"required":["updatedAt"]},"CreateLinkDto":{"type":"object","properties":{"url":{"type":"string","description":"url"},"cateId":{"type":"string","description":"category id"}},"required":["url"]},"UpdateLinkDto":{"type":"object","properties":{"url":{"type":"string","description":"url"},"name":{"type":"string","maxLength":80,"description":"name"},"title":{"type":"string","maxLength":150,"description":"title"},"description":{"type":"string","maxLength":500,"description":"description"},"icon":{"type":"string","maxLength":500,"description":"icon"},"weight":{"type":"number","description":"weight"},"cateId":{"type":"object","description":"Category"}},"required":["url","name","title","description","icon","cateId"]}}}}